requirements-ingest.txt
google-cloud-firestore>=2.15.0

